requirements.txt
# CLOB V2 SDK (V1 package no longer works on production).
py-clob-client-v2
websockets>=12.0
python-dotenv>=1.0.0
flask>=3.0.0
